@@ -1,12 +1,13 @@
 cmake_minimum_required(VERSION 3.26)
-project(CMake_Template CXX)
+project(Zappy CXX C)
 
-set(CMAKE_CXX_STANDARD 20)
-set(CMAKE_CXX_STANDARD_REQUIRED YES)
-set(CMAKE_CXX_FLAGS "-Wall -Wextra")
+set(CMAKE_C_STANDARD 17)
+set(CMAKE_C_STANDARD_REQUIRED YES)
 set(CMAKE_C_FLAGS "-Wall -Wextra")
 
-set(EXECUTABLE_NAME template)
+set(CMAKE_CXX_STANDARD 23)
+set(CMAKE_CXX_STANDARD_REQUIRED YES)
+set(CMAKE_CXX_FLAGS "-Wall -Wextra")
 
 add_library(build_configs INTERFACE)
 
@@ -14,7 +15,6 @@ target_compile_options(build_configs INTERFACE
     $<$<CONFIG:DEBUG>:-ggdb3>
     $<$<CONFIG:DEBUG>:-Og>
     $<$<CONFIG:RELEASE>:-Ofast>
-    -march=native
     -flto=auto
 )
 
@@ -30,23 +30,6 @@ target_link_options(build_configs INTERFACE
     -flto=auto
 )
 
-
-add_executable(template
-    src/main.cpp
-)
-
-target_include_directories(template PRIVATE
-    include
-)
-
-target_link_libraries(template PRIVATE
-    build_configs
-)
-
-set_target_properties(template PROPERTIES
-    OUTPUT_NAME ${EXECUTABLE_NAME}
-)
-
-add_custom_command(TARGET template
-    POST_BUILD COMMAND cp ${EXECUTABLE_NAME} ../
-)
+add_subdirectory(gui)
+add_subdirectory(server)
+add_subdirectory(ai)
