@@ -1,7 +1,7 @@
-#include <cstdio>
+#include <print>
 
 int main(int argc, char *argv[])
 {
-    printf("Hello World\n");
+    std::print("Hello World\n");
     return 0;
 }
